ipv6支持简写和全写的精确查询
